[PKGBUG] Fix issues from moving code to unlmod
