+++ JavaEFS/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>JavaEFS0</name>
+	<name>JavaEFS</name>
 	<comment></comment>
 	<projects>
 	</projects>
