+++ JavaEFS/unlmod/edu/unl/cse/guitarext/GUITARDataFiles.java
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -28,11 +28,12 @@ import javax.xml.bind.Unmarshaller;
 
 import edu.umd.cs.guitar.model.data.*;
 /**
- * 
+ *
  * @author Jonathan A. Saddler
  *
  */
 public class GUITARDataFiles {
+	@SuppressWarnings("serial")
 	public static class WrongFileTypeException extends IllegalArgumentException
 	{
 		public final String expected, read;
